SchottkyAnoMaLy

Documentation for SchottkyAnoMaLy.

SchottkyAnoMaLy.DonutVolcanoEnsemble

SchottkyAnoMaLy.DonutVolcanoEnsemble

SchottkyAnoMaLy.DonutVolcanoEnsemble

Base.append!

Base.eltype

Base.push!

SchottkyAnoMaLy.Theta

SchottkyAnoMaLy.donutvolcano

SchottkyAnoMaLy.ensemble

SchottkyAnoMaLy.get_pair

SchottkyAnoMaLy.norm_donutvolcano

SchottkyAnoMaLy.npairs

SchottkyAnoMaLy.valid

SchottkyAnoMaLy.μvalue

SchottkyAnoMaLy.σvalue

SchottkyAnoMaLy.DonutVolcanoEnsemble — Type

DonutVolcanoEnsemble{T <: Real}

An ensemble of $M$ donutvolcanos takes the form

\[\mathrm{p}_M(x) = \frac{1}{M} \sum_{k=1}^M \mathrm{donutvolcano}(x, \mu_k, \sigma_k).\]

This object is a wrapper around an ensemble::Vector{Tuple{T, T}} for the individual donutvolcanos. The values of $(μ_k, σ_k)$ are restricted such that $\mu_k \geq 0$ and $\sigma_k > 0$ via the valid function.

SchottkyAnoMaLy.DonutVolcanoEnsemble — Method

(::DonutVolcanoEnsemble)(x)

Evaluate the DonutVolcanoEnsemble as a function over a set of values x.

julia> dve = DonutVolcanoEnsemble(Float64, [(0, 1)]);

julia> dve(2)
0.2706705664732254

julia> dve([3, 4])
2-element Vector{Float64}:
 0.033326989614726917
 0.0013418505116100474

SchottkyAnoMaLy.DonutVolcanoEnsemble — Method

DonutVolcanoEnsemble(args...)

Various interfaces to a DonutVolcanoEnsemble

Copy/Move constructors

DonutVolcanoEnsemble(::Vector{Tuple{T, T}})

Create an ensemble from a Vector of (μ, σ) Tuples of the same Type.

DonutVolcanoEnsemble(::Vector{Tuple{T, U}})

Create an ensemble from a Vector of (μ, σ) Tuples of the different Types.

The type of the ensemble is chosen by Base.promote_type.

DonutVolcanoEnsemble(::Type{S}, ::Vector{Tuple{T, U}})

Create an ensemble from a Vector of (μ, σ) Tuples of the different Types into an ensemble of with preferred type S.

The ultimate type of the ensemble is chosen by Base.promote_type.

Note

The constructors above will check that any supplied pair is valid.

Empty constructors

DonutVolcanoEnsemble(::Type{S})

Create an empty ensemble of chosen type S.

DonutVolcanoEnsemble()

Create an empty ensemble. By default the type is Float64.

Base.append! — Method

Base.append!(::DonutVolcanoEnsemble, vals)

Append the vals::Vector{Tuple{S, T}} to the DonutVolcanoEnsemble.

julia> dve = DonutVolcanoEnsemble()
DonutVolcanoEnsemble{Float64}(Tuple{Float64, Float64}[])

julia> append!(dve, [(0, 1), (1, 3)])
2-element Vector{Tuple{Float64, Float64}}:
 (0.0, 1.0)
 (1.0, 3.0)

julia> dve
DonutVolcanoEnsemble{Float64}([(0.0, 1.0), (1.0, 3.0)])

Note

This function will check that the pair is valid.

Base.eltype — Method

Base.eltype(::DonutVolcanoEnsemble)

Return the type of the DonutVolcanoEnsemble.

julia> dve = DonutVolcanoEnsemble()
DonutVolcanoEnsemble{Float64}(Tuple{Float64, Float64}[])

julia> eltype(dve)
Float64

Base.push! — Method

Base.push!(::DonutVolcanoEnsemble, pair)

Push a new (μ, σ) Tuple into a given DonutVolcanoEnsemble.

julia> dve = DonutVolcanoEnsemble()
DonutVolcanoEnsemble{Float64}(Tuple{Float64, Float64}[])

julia> push!(dve, (0, 1))
1-element Vector{Tuple{Float64, Float64}}:
 (0.0, 1.0)

julia> dve
DonutVolcanoEnsemble{Float64}([(0.0, 1.0)])

Note

This function will check that the pair is valid.

SchottkyAnoMaLy.Theta — Method

Theta(x)

Package-specific implementation of the Heaviside Theta function. Our definition is

\[\Theta(x) = \begin{cases} 1, & x \geq 0 \\ 0, & \mathrm{otherwise} \end{cases}.\]

SchottkyAnoMaLy.donutvolcano — Method

donutvolcano(x, μ, σ)

The donutvolcano distribution calculated for an input x with center μ and width σ. The argument x can be a scalar or broadcastable.

The distribution itself is given by

\[{\rm donutvolcano}(x; μ, σ) = \frac{x \Theta(x)}{\mathcal{N}(\mu,\sigma)} \exp\left[ -\frac{\left(x - \mu\right)^2}{2\sigma^2} \right],\]

where we use our own Heaviside Theta function above. The normalization constant $\mathcal{N}(\mu,\sigma)$ given by norm_donutvolcano.

SchottkyAnoMaLy.ensemble — Method

ensemble(::DonutVolcanoEnsemble)

Getter for the Vector{Tuple{T, T}} of the given ensemble.

SchottkyAnoMaLy.get_pair — Method

get_pair(::DonutVolcanoEnsemble, pair_idx) -> Tuple{T, T}

Getter for a single pair (Tuple{T, T}) at a specified pair_idx.

SchottkyAnoMaLy.norm_donutvolcano — Method

norm_donutvolcano(μ, σ)

The normalization for a donutvolcano distribution with center μ and width σ. The explicit form is given by

\[\mathcal{N}(\mu, \sigma) = \sqrt{\frac{\pi }{2}} \mu \sigma \left[\text{erf}\left(\frac{\mu }{\sqrt{2} \sigma }\right)+1\right]+\sigma ^2 {\rm e}^{-\frac{\mu ^2}{2 \sigma ^2}}.\]

SchottkyAnoMaLy.npairs — Method

npairs(::DonutVolcanoEnsemble)

Return the number of donutvolcano in the DonutVolcanoEnsemble.

SchottkyAnoMaLy.valid — Method

valid(::Tuple{T, T})

Returns the given Tuple (μ, σ) if μ ≥ 0 and σ > 0. Otherwise, an AssertionError will be thrown.

SchottkyAnoMaLy.μvalue — Method

μvalue(::Tuple{T, T})
μvalue(::DonutVolcanoEnsemble, idx)

Return the value of μ from a given (μ, σ) Tuple.

SchottkyAnoMaLy.σvalue — Method

σvalue(::Tuple{T, T})
σvalue(::DonutVolcanoEnsemble, idx)

Return the value of σ from a given (μ, σ) Tuple.