SchottkyanoMaLy

Documentation for SchottkyanoMaLy.

SchottkyanoMaLy.DonutVolcanoEnsemble

SchottkyanoMaLy.DonutVolcanoEnsemble

SchottkyanoMaLy.Theta

SchottkyanoMaLy.donutvolcano

SchottkyanoMaLy.ensemble

SchottkyanoMaLy.get_pair

SchottkyanoMaLy.norm_donutvolcano

SchottkyanoMaLy.npairs

SchottkyanoMaLy.DonutVolcanoEnsemble — Type

DonutVolcanoEnsemble{T <: AbstractFloat}

An ensemble of $M$ donutvolcanos takes the form

\[\mathrm{p}_M(x) = \frac{1}{M} \sum_{k=1}^M \mathrm{donutvolcano}(x, \mu_k, \sigma_k).\]

This object is a wrapper around an ensemble::Vector{Tuple{T, T}} for the individual donutvolcanos.

SchottkyanoMaLy.DonutVolcanoEnsemble — Method

DonutVolcanoEnsemble(args...)

Various interfaces to a DonutVolcanoEnsemble

Copy/Move constructors

DonutVolcanoEnsemble(::Vector{Tuple{T, T}})

Create an ensemble from a Vector of (μ, σ) Tuples of the same Type.

DonutVolcanoEnsemble(::Vector{Tuple{T, U}})

Create an ensemble from a Vector of (μ, σ) Tuples of the different Types.

The type of the ensemble is chosen by Base.promote_type.

DonutVolcanoEnsemble(::Type{S}, ::Vector{Tuple{T, U}})

Create an ensemble from a Vector of (μ, σ) Tuples of the different Types into an ensemble of with preferred type S.

The ultimate type of the ensemble is chosen by Base.promote_type.

Empty constructors

DonutVolcanoEnsemble(::Type{S})

Create an empty ensemble of chosen type S.

DonutVolcanoEnsemble()

Create an empty ensemble. By default the type is Float64.

SchottkyanoMaLy.Theta — Method

Theta(x)

Package-specific implementation of the Heaviside Theta function. Our definition is

\[\Theta(x) = \begin{cases} 1, & x \geq 0 \\ 0, & \mathrm{otherwise} \end{cases}.\]

SchottkyanoMaLy.donutvolcano — Method

donutvolcano(x, μ, σ)

The donutvolcano distribution calculated for an input x with center μ and width σ. The argument x can be a scalar or broadcastable.

The distribution itself is given by

\[{\rm donutvolcano}(x; μ, σ) = \frac{x \Theta(x)}{\mathcal{N}(\mu,\sigma)} \exp\left[ -\frac{\left(x - \mu\right)^2}{2\sigma^2} \right],\]

where we use our own Heaviside Theta function above. The normalization constant $\mathcal{N}(\mu,\sigma)$ given by norm_donutvolcano.

SchottkyanoMaLy.ensemble — Method

ensemble(::DonutVolcanoEnsemble)

Getter for the Vector{Tuple{T, T}} of the given ensemble.

SchottkyanoMaLy.get_pair — Method

get_pair(::DonutVolcanoEnsemble, pair_idx) -> Tuple{T, T}

Getter for a single pair (Tuple{T, T}) at a specified pair_idx.

SchottkyanoMaLy.norm_donutvolcano — Method

norm_donutvolcano(μ, σ)

The normalization for a donutvolcano distribution with center μ and width σ. The explicit form is given by

\[\mathcal{N}(\mu, \sigma) = \sqrt{\frac{\pi }{2}} \mu \sigma \left[\text{erf}\left(\frac{\mu }{\sqrt{2} \sigma }\right)+1\right]+\sigma ^2 {\rm e}^{-\frac{\mu ^2}{2 \sigma ^2}}.\]

SchottkyanoMaLy.npairs — Method

npairs(::DonutVolcanoEnsemble)

Return the number of donutvolcano in the DonutVolcanoEnsemble.