SchottkyAnoMaLy

Documentation for SchottkyAnoMaLy.

SchottkyAnoMaLy.DonutVolcanoEnsemble

SchottkyAnoMaLy.DonutVolcanoEnsemble

SchottkyAnoMaLy.DonutVolcanoEnsemble

SchottkyAnoMaLy.NLevelSystem

SchottkyAnoMaLy.RandomDonutVolcanoGenerator

SchottkyAnoMaLy.TwoLevelSystem

Base.append!

Base.eltype

Base.length

Base.push!

Base.rand

Base.rand

SchottkyAnoMaLy.Theta

SchottkyAnoMaLy.chebycoefficients

SchottkyAnoMaLy.chebycoefficients!

SchottkyAnoMaLy.chebydecomposition

SchottkyAnoMaLy.compute_cVs

SchottkyAnoMaLy.compute_cVs!

SchottkyAnoMaLy.create_chebinterp

SchottkyAnoMaLy.donutvolcano

SchottkyAnoMaLy.ensemble

SchottkyAnoMaLy.gausskernel

SchottkyAnoMaLy.get_pair

SchottkyAnoMaLy.input_test_msqdiffs

SchottkyAnoMaLy.input_test_msqdiffs!

SchottkyAnoMaLy.kernel

SchottkyAnoMaLy.mass2

SchottkyAnoMaLy.minimizing_component

SchottkyAnoMaLy.minimizing_solution

SchottkyAnoMaLy.msqdiff

SchottkyAnoMaLy.norm_donutvolcano

SchottkyAnoMaLy.populate_msqdiffs

SchottkyAnoMaLy.populate_msqdiffs!

SchottkyAnoMaLy.specific_heat

SchottkyAnoMaLy.specific_heat

SchottkyAnoMaLy.specific_heat

SchottkyAnoMaLy.valid

SchottkyAnoMaLy.μvalue

SchottkyAnoMaLy.σvalue

SchottkyAnoMaLy.∂σ_gausskernel

SchottkyAnoMaLy.DonutVolcanoEnsemble — Type

DonutVolcanoEnsemble{T <: Real}

An ensemble of $M$ donutvolcanos takes the form

\[\mathrm{p}_M(x) = \frac{1}{M} \sum_{k=1}^M \mathrm{donutvolcano}(x, \mu_k, \sigma_k).\]

This object is a wrapper around an ensemble::Vector{Tuple{T, T}} for the individual donutvolcanos. The values of $(μ_k, σ_k)$ are restricted such that $\mu_k \geq 0$ and $\sigma_k > 0$ via the valid function.

SchottkyAnoMaLy.DonutVolcanoEnsemble — Method

(::DonutVolcanoEnsemble)(x)

Evaluate the DonutVolcanoEnsemble as a function over a set of values x.

julia> dve = DonutVolcanoEnsemble(Float64, [(0, 1)]);

julia> dve(2)
0.2706705664732254

julia> dve([3, 4])
2-element Vector{Float64}:
 0.033326989614726917
 0.0013418505116100474

SchottkyAnoMaLy.DonutVolcanoEnsemble — Method

DonutVolcanoEnsemble(args...)

Various interfaces to a DonutVolcanoEnsemble

Copy/Move constructors

DonutVolcanoEnsemble(::Vector{Tuple{T, T}})

Create an ensemble from a Vector of (μ, σ) Tuples of the same Type.

DonutVolcanoEnsemble(::Vector{Tuple{T, U}})

Create an ensemble from a Vector of (μ, σ) Tuples of the different Types.

The type of the ensemble is chosen by Base.promote_type.

DonutVolcanoEnsemble(::Type{S}, ::Vector{Tuple{T, U}})

Create an ensemble from a Vector of (μ, σ) Tuples of the different Types into an ensemble of with preferred type S.

The ultimate type of the ensemble is chosen by Base.promote_type.

Note

The constructors above will check that any supplied pair is valid.

Empty constructors

DonutVolcanoEnsemble(::Type{S})

Create an empty ensemble of chosen type S.

DonutVolcanoEnsemble()

Create an empty ensemble. By default the type is Float64.

SchottkyAnoMaLy.NLevelSystem — Type

abstract type NLevelSystem  end

Create a supertype for all non-interacting systems with discrete energy levels.

Each new concrete subtype must define a specific_heat function.

SchottkyAnoMaLy.RandomDonutVolcanoGenerator — Type

RandomDonutVolcanoGenerator{T <: Real}

Object to generate a random DonutVolcanoEnsemble.

SchottkyAnoMaLy.TwoLevelSystem — Type

struct TwoLevelSystem end

Empty struct for the case of the two-level system. Subtype of NLevelSystem. Used to define traits for this specific type of system.

Base.append! — Method

Base.append!(::DonutVolcanoEnsemble, vals)

Append the vals::Vector{Tuple{S, T}} to the DonutVolcanoEnsemble.

julia> dve = DonutVolcanoEnsemble()
DonutVolcanoEnsemble{Float64}(Tuple{Float64, Float64}[])

julia> append!(dve, [(0, 1), (1, 3)])
2-element Vector{Tuple{Float64, Float64}}:
 (0.0, 1.0)
 (1.0, 3.0)

julia> dve
DonutVolcanoEnsemble{Float64}([(0.0, 1.0), (1.0, 3.0)])

Note

This function will check that the pair is valid.

Base.eltype — Method

Base.eltype(::DonutVolcanoEnsemble)

Return the type of the DonutVolcanoEnsemble.

julia> dve = DonutVolcanoEnsemble()
DonutVolcanoEnsemble{Float64}(Tuple{Float64, Float64}[])

julia> eltype(dve)
Float64

Base.length — Method

length(::DonutVolcanoEnsemble)

Return the number of donutvolcano in the DonutVolcanoEnsemble.

Base.push! — Method

Base.push!(::DonutVolcanoEnsemble, pair)

Push a new (μ, σ) Tuple into a given DonutVolcanoEnsemble.

julia> dve = DonutVolcanoEnsemble()
DonutVolcanoEnsemble{Float64}(Tuple{Float64, Float64}[])

julia> push!(dve, (0, 1))
1-element Vector{Tuple{Float64, Float64}}:
 (0.0, 1.0)

julia> dve
DonutVolcanoEnsemble{Float64}([(0.0, 1.0)])

Note

This function will check that the pair is valid.

Base.rand — Method

rand([rng = GLOBAL_RNG], ::RandomDonutVolcanoGenerator{T}, num_ensembles::Int)

Return a Vector of random DonutVolcanoEnsembles of length num_ensembles. The type of each DonutVolcanoEnsemble is T. By default, the GLOBAL_RNG is used.

Base.rand — Method

rand([rng = GLOBAL_RNG], ::RandomDonutVolcanoGenerator{T})

Generate a random DonutVolcanoEnsemble of type T based on the RandomDonutVolcanoGenerator supplied. By default, T = Float64.

julia> using Random 

julia> rng = MersenneTwister(42);  # Choice for longevity. Use Xoshiro in practice or the GLOBAL_RNG.

julia> rdveg = RandomDonutVolcanoGenerator(3, 10, 10);

julia> dve = rand(rng, rdveg);     # Suppress output for floating-point error reasons.

julia> for tup ∈ ensemble(dve)
           @show round.( tup; digits = 6 )
       end
round.(tup; digits = 6) = (6.23099, 2.780566)
round.(tup; digits = 6) = (9.745867, 4.499406)
round.(tup; digits = 6) = (8.427712, 3.660781)

SchottkyAnoMaLy.Theta — Method

Theta(x)

Package-specific implementation of the Heaviside Theta function. Our definition is

\[\Theta(x) = \begin{cases} 1, & x \geq 0 \\ 0, & \mathrm{otherwise} \end{cases}.\]

SchottkyAnoMaLy.chebycoefficients — Function

chebycoefficients(ensembles, order, Δmin, Δmax, [coefftol = zero(Float64)])

Allocate the appropriate coefficients Matrix and pass it to chebycoefficients!.

SchottkyAnoMaLy.chebycoefficients! — Function

chebycoefficients!(coefficients, ensembles, order, Δmin, Δmax, [coefftol = zero(Float64)])

Compute all of the Chebyshev coefficients for each of the ensembles to a given order on the interval [Δmin, Δmax]. This function uses chebydecomposition to do so, and assumes the coefficients argument is a length(ensembles) × (order + 1) Matrix.

Because this is a somewhat complex calculation, it should probably be calculated once.

SchottkyAnoMaLy.chebydecomposition — Function

chebydecomposition(ensemble, order, Δmin, Δmax, [ coefftol = zero(Float64) ])

Use FastChebInterp.jl to compute the Chebyshev interpolation of a given ensemble to a polynomial order on the [Δmin, Δmax] interval. This function returns the coefficients $a_k$ computed by FastChebInterp.chebinterp and defined by

\[\mathrm{p}(\Delta) \approx \sum_{k=0}^M a_k T_k(\Delta),\]

where $\mathrm{p}(\Delta)$ is the ensemble, $M$ is the order, and $T_k(\Delta)$ is the $k^{\mathrm{th}}$ Chebyshev polynomial of the first kind.

Note

Changing the optional coefftol from the default value of 0.0 may lead to issues downstream. Essentially, FastChebInterp.jl uses this (I think) for precision purposes, but if it is nonzero then the output order is not guaranteed to be what is provided as and argument.

SchottkyAnoMaLy.compute_cVs! — Method

compute_cVs!( cV_arr, ensembles, temps; [kwargs...] )

Compute the specific_heat at each temperature in temps for a given ensemble as a single column. By assumption, cV_arr is a length(temps) × length(ensembles) Matrix, and it has been initialized with zeros.

Keyword arguments

nls =TwoLevelSystem(): The type of NLevelSystem used to compute the specific_heat.

Δmin = 0.: The minimum value of the level-splitting in the integration over each ensemble.

Δmax = Inf: The maximum value of the level-splitting in the integration over each ensemble.

rtol = sqrt(eps()): The relative tolerance used in QuadGK.quadgk to integrate over the ensembles.

Note

Since this populates the specific heat values with QuadGK.quadgk, it is best only to compute once as it's a bit of an expensive operation.

SchottkyAnoMaLy.compute_cVs — Method

compute_cVs(ensembles, temps; [kwargs...])

Creates a zero Matrix of size length(temps) × length(ensembles) and passes it to compute_cVs!.

SchottkyAnoMaLy.create_chebinterp — Function

create_chebinterp(coefficients, lowerbound, upperbound, [N = 1])

Create a new FastChebInterp.ChebPoly from a set of Chebyshev coefficients. By default (and until I can think of a reason to generalize), this will create a 1D ChebPoly. By assumption, the lowerbound and upperbound are numbers.

SchottkyAnoMaLy.donutvolcano — Method

donutvolcano(x, μ, σ)

The donutvolcano distribution calculated for an input x with center μ and width σ. The argument x can be a scalar or broadcastable.

The distribution itself is given by

\[{\rm donutvolcano}(x; μ, σ) = \frac{x \Theta(x)}{\mathcal{N}(\mu,\sigma)} \exp\left[ -\frac{\left(x - \mu\right)^2}{2\sigma^2} \right],\]

where we use our own Heaviside Theta function above. The normalization constant $\mathcal{N}(\mu,\sigma)$ given by norm_donutvolcano.

SchottkyAnoMaLy.ensemble — Method

ensemble(::DonutVolcanoEnsemble)

Getter for the Vector{Tuple{T, T}} of the given ensemble.

SchottkyAnoMaLy.gausskernel — Method

gausskernel(d2, hypσ, [method = TrapezoidalFast()])
gausskernel(xdata, y1data, y2data, hypσ, [method = TrapezoidalFast()])

Compute the Gaussian kernel function for two sets of ydata, given the hyperparameter hypσ. This kernel function is of the form

\[K[y_1(x), y_2(x); \sigma] = \exp\left[ - \frac{d^2[y_1(x), y_2(x)]}{2\sigma^2} \right],\]

where $d^2$ is the msqdiff functional.

SchottkyAnoMaLy.get_pair — Method

get_pair(::DonutVolcanoEnsemble, pair_idx) -> Tuple{T, T}

Getter for a single pair (Tuple{T, T}) at a specified pair_idx.

SchottkyAnoMaLy.input_test_msqdiffs! — Method

input_test_msqdiffs!(output, test_cV, cV_arr, temps, [method = TrapezoidalFast()])

Compute a single column vector of the msqdiffs between the test_cV argument and each column of the cV_arr, all measured as functions of temps temperature. Store the result in the output vector.

SchottkyAnoMaLy.input_test_msqdiffs — Method

input_test_msqdiffs(test_cV, cV_arr, temps, [method = TrapezoidalFast()])

Allocate an appropriate output Vector and pass it to input_test_msqdiffs!.

SchottkyAnoMaLy.kernel — Method

kernel(msqdiff_matrix, hypσ)

Compute the element-wise gausskernel of the msqdiff_matrix with respect to the hyperparameter hypσ. This makes a copy.

SchottkyAnoMaLy.mass2 — Function

mass2( xdata, ydata, [method = TrapezoidalFast()] )

Use NumericalIntegration.jl to compute the "mean-square-mass" of a set of data ydata over the domain xdata.

Note

The default IntegrationMethod is set to TrapezoidalFast which will lead to errors if xdata and ydata are of different lengths. For experimental data inputs, however, this should be fine because the specific heat is measured for a specific temperature.

The mean-square-mass formula, given as a continuous integral, is

\[m^2[y(x)] = \frac{1}{b - a}\int_a^b \mathrm{d}x\, \left[ y(x) \right]^2.\]

julia> xvals = LinRange(1, 10, 10);

julia> yvals = sqrt.(xvals);  # In this case the trapezoidal rule is exact

julia> value = SchottkyAnoMaLy.mass2(xvals, yvals)
5.5

julia> value ≈ 1/(10 - 1) * 1/2 * ( maximum(xvals)^2 - minimum(xvals)^2 )
true

SchottkyAnoMaLy.minimizing_component — Method

minimizing_component(kernel_column, inv_kernel_matrix, train_component_vector)

Compute the solution to the kernel ridge regression problem (KRR). This solution minimizes the regularized square-error

\[\mathcal{E}_m = \sum_i \left( f^m_i - \sum_j K(i, j; \sigma) \alpha^m_j \right)^2 + \lambda \sum_{ij} \alpha_i K(i, J) \alpha_j,\]

where $K(i, j)$ is the kernel function measured between functions $i$ and $j$, $f^m_i$ is the $m^{\mathrm{th}}$ component of $i^{\mathrm{th}}$ training function, $\sigma$ and $\lambda$ are the hyperparameters, and $\boldsymbol{\alpha}^m$ is related to the $\boldsymbol{f}^m$ training vector as

\[\boldsymbol{\alpha}^m = \left( K(i,j;\sigma) + \lambda \mathbb{1} \right)^{-1} \boldsymbol{f}^m.\]

The prediction that minimizes $\mathcal{E}_m$ is given by

\[f^m_{\mathrm{out}} = \boldsymbol{K}^T_{in} \boldsymbol{\alpha}^m.\]

By direct substitution, this function then computes

\[f^m_{\mathrm{out}} = \boldsymbol{K}^T_{in} \left( K(i,j;\sigma) + \lambda \mathbb{1} \right)^{-1} \boldsymbol{f}^m,\]

where $\boldsymbol{K}^T_{in}$ = kernel_column, the whole inverse is inv_kernel_matrix, and $\boldsymbol{f}^m$ = train_component_vector.

SchottkyAnoMaLy.minimizing_solution — Method

minimizing_solution(kernel_column, inv_kernel_matrix, train_component_matrix)

Use minimizing_component to compute the full solution to the Kernel Ridge Regression problem. This assumes that train_component_matrix is of size N × M, where N is the number of elements in the training set and M is the number of Chebyshev coefficients (order + 1).

SchottkyAnoMaLy.msqdiff — Function

msqdiff(xdata, y1data, y2data, [method = TrapezoidalFast()])

Compute the mean-square-difference between two sets of ydata on the given xdata domain. This function calls mass2 to perform the integration.

The formula for the mean-square-difference between functions $y_1$ and $y_2$ on the continuous interval (a,b) is given by

\[d^2(y_1, y_2) = m^2[y_1(x) - y_2(x)] = \frac{1}{b - a}\int_a^b \mathrm{d}x\, \left[ y_1(x) - y_2(x) \right]^2.\]

SchottkyAnoMaLy.norm_donutvolcano — Method

norm_donutvolcano(μ, σ)

The normalization for a donutvolcano distribution with center μ and width σ. The explicit form is given by

\[\mathcal{N}(\mu, \sigma) = \sqrt{\frac{\pi }{2}} \mu \sigma \left[\text{erf}\left(\frac{\mu }{\sqrt{2} \sigma }\right)+1\right]+\sigma ^2 {\rm e}^{-\frac{\mu ^2}{2 \sigma ^2}}.\]

SchottkyAnoMaLy.populate_msqdiffs! — Method

populate_msqdiffs!( msqdiff_mat, cV_arr, temps; [kwargs...])

Compute the msqdiff for each pair of specific heats in cV_arr as a function of temperature (temps) as the xdata. This assumes that

msqdiff_mat is a N × N zero Matrix,

N == size(cV_arr)[2], where cV_arr is a specific heat Matrix calculated from compute_cVs!,

the number of temperatures in temps equals the number of specific heat rows.

This function uses input_test_msqdiffs! to fill the Matrix. The input is the specific heat of a given ensemble, and the tests are those from all other ensembles.

Note

N is likely to be a large number, so it is best to use this function once.

SchottkyAnoMaLy.populate_msqdiffs — Method

populate_msqdiffs(cV_arr, temps; [kwargs...])

Creates a zero Matrix of size length(ensembles) × length(ensembles) and passes it to populate_msqdiffs!.

SchottkyAnoMaLy.specific_heat — Function

specific_heat([::NLevelSystem = TwoLevelSystem()], T, ::DonutVolcanoEnsemble, [Δmin = 0, Δmax = Inf])

Calculate the specific_heat $\tilde{c}_V(T, \Delta$ integrated over the DonutVolcanoEnsemble for a fixed temperature $T$. By default, this is calculated for a TwoLevelSystem.

The formula for this integration is given by

\[c_V(T) = \int_{\Delta_{\min}}^{\Delta_{\max}} \mathrm{d}\Delta\, \mathrm{p}(\Delta) \tilde{c}_V(T, \Delta),\]

where $\mathrm{p}(\Delta)$ is calculated from the DonutVolcanoEnsemble.

SchottkyAnoMaLy.specific_heat — Method

specific_heat(T, Δ) = specific_heat(TwoLevelSystem, T, Δ)

The default specific heat implementation chooses a TwoLevelSystem.

julia> specific_heat(1, 1)
0.41997434161402614

julia> specific_heat(TwoLevelSystem(), 1, 1)
0.41997434161402614

SchottkyAnoMaLy.specific_heat — Method

specific_heat(::TwoLevelSystem, T, Δ)

Define the specific heat function for a two-level system at temperature T and level-splitting Δ (measured in units of temperature). The formula is given by

\[c_V(T;\, \Delta) = \left( \frac{\Delta}{T} \right)^2 \mathrm{sech}^2 \left( \frac{\Delta}{T} \right)^2.\]

julia> specific_heat(TwoLevelSystem(), 1, 1)
0.41997434161402614

SchottkyAnoMaLy.valid — Method

valid(::Tuple{T, T})

Returns the given Tuple (μ, σ) if μ ≥ 0 and σ > 0. Otherwise, an AssertionError will be thrown.

SchottkyAnoMaLy.μvalue — Method

μvalue(::Tuple{T, T})
μvalue(::DonutVolcanoEnsemble, idx)

Return the value of μ from a given (μ, σ) Tuple.

SchottkyAnoMaLy.σvalue — Method

σvalue(::Tuple{T, T})
σvalue(::DonutVolcanoEnsemble, idx)

Return the value of σ from a given (μ, σ) Tuple.

SchottkyAnoMaLy.∂σ_gausskernel — Method

∂σ_gausskernel(d2, hypσ)  # \partial<TAB>\sigma<TAB>
∂σ_gausskernel(d2, gK, hypσ)

Calculate the gradient of gausskernel with respect to the hyperparameter hypσ. The explicit formula is given by

\[\partial_\sigma K(d^2; \sigma) = \sigma^{-3} d^2 K(d^2; \sigma).\]

where the $K(d^2; \sigma)$ function is gausskernel and gK in the second method for this function.